feat(billing): support fractional pricing and improve UX
- Change pricing unit prices from int64 to float64/DECIMAL for fractional pricing
- Update billing_service to convert dollar prices to cents for Stripe
- Fix infinite loading in analytics page with stable query keys
- Optimize analytics queries to fetch only needed columns
- Remove warning alerts from end-users and pricing-models pages
- Polish UX with empty states and loading indicators

+++ svc/api/routes/v1_billing_pricing_models_create/handler.go
@@ -22,8 +22,8 @@ import (
 type CreatePricingModelRequest struct {
 	Name                  string                 `json:"name" validate:"required"`
 	Currency              string                 `json:"currency" validate:"required,len=3"`
-	VerificationUnitPrice int64                  `json:"verificationUnitPrice" validate:"gte=0"`
-	RatelimitUnitPrice    int64                  `json:"ratelimitUnitPrice" validate:"gte=0"`
+	VerificationUnitPrice float64                `json:"verificationUnitPrice" validate:"gte=0"`
+	RatelimitUnitPrice    float64                `json:"ratelimitUnitPrice" validate:"gte=0"`
 	TieredPricing         *billing.TieredPricing `json:"tieredPricing,omitempty"`
 }
 
@@ -36,8 +36,8 @@ type CreatePricingModelResponse struct {
 		WorkspaceID           string                 `json:"workspaceId"`
 		Name                  string                 `json:"name"`
 		Currency              string                 `json:"currency"`
-		VerificationUnitPrice int64                  `json:"verificationUnitPrice"`
-		RatelimitUnitPrice    int64                  `json:"ratelimitUnitPrice"`
+		VerificationUnitPrice float64                `json:"verificationUnitPrice"`
+		RatelimitUnitPrice    float64                `json:"ratelimitUnitPrice"`
 		TieredPricing         *billing.TieredPricing `json:"tieredPricing,omitempty"`
 		Version               int32                  `json:"version"`
 		Active                bool                   `json:"active"`
